Release CPython89 v3.5.3
This a maintenance release of the ‘CPython89’ project, which provides
ANSI X3.159-1989 soure code compatibility for ‘CPython 3.5.3’.

With this release, the CPython project adds a ‘pyshellext.vcxproj’
(A shell extension implementing a Drag'n'Drop handler for data files
dropped on Python scripts).

An attempted backport of this project hasn't been successful – for the
simple reason, that its implementation relies on the WRL (Windows
Runtime C++ Template Library), which isn't available in Visual C++ 9.0.